rename cluster setting
